Don't compile against Apple Secure Transport
We want a version of curl that's only compiled against openssl.

pycurl 7.43.0.5 doesn't work with libcurl when it's compiled against
this as well as openssl - it can't detect which SSL backend libcurl is
using and raises an error at runtime.

pycurl 7.43.0.6 fixes this issue, but we can't upgrade to that version
until celery[sqs] has been updated to allow that version (it currently
has an exact dependency on 7.43.0.5 of pycurl)